customer-support-agent-sql/requirements.txt
htbuilder>=0.9.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
pyodbc>=5.0.0
streamlit>=1.51.0
tqdm>=4.66.0